ZMX-#55 renamed inputHandler to Cli (input interface)
